+++ SMFL/LerpMoveTo.cpp
@@ -5,7 +5,7 @@ LerpMoveTo::LerpMoveTo(float speed) : speed(speed)
 {
 }
 
-void LerpMoveTo::MoveTo(GameObject& object, sf::Vector2f destination, double time)
+void LerpMoveTo::MoveTo(GameObject& object, sf::Vector2f destination, float time)
 {
 	object.SetPosition(VectorUtility::Lerp(object.GetPosition(), destination, time*speed));
 }
