+++ SMFL/SpringMoveTo.h
@@ -7,7 +7,7 @@ class SpringMoveTo : public IMoveTo
 public:
 	SpringMoveTo(float constant, float damping);
 	virtual ~SpringMoveTo() {};
-	virtual void MoveTo(GameObject& object, sf::Vector2f destination, double time);
+	virtual void MoveTo(GameObject& object, sf::Vector2f destination, float time);
 private:
 	float constant;
 	float damping;
